// src/lib/clients/myanimelist/generated/models/UserMangaListEdge.ts
/* generated using openapi-typescript-codegen -- do no edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */

import type { MangaForList } from './MangaForList';
import type { MangaListStatus } from './MangaListStatus';

export type UserMangaListEdge = {
    node?: MangaForList;
    list_status?: MangaListStatus;
};